pyproject.toml
[project]
name = "oscribe"
version = "0.1.0"
description = "Fast speech-to-text with system tray integration"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "faster-whisper>=1.1.0",
    "onnxruntime>=1.20,<1.24",
    "numpy>=2.0.0",
    "pyperclip>=1.9.0",
    "pyqt6>=6.7.0",
    "pyzmq>=26.0.0",
    "pyyaml>=6.0",
    "sounddevice>=0.5.0",
]

[project.optional-dependencies]
cuda = [
    "nvidia-cublas-cu12",
    "nvidia-cudnn-cu12",
]
rocm = []  # ROCm CTranslate2 wheel installed by install.sh from GitHub releases

[project.scripts]
oscribe = "oscribe.service:main"
oscribe-trigger = "oscribe.trigger:main"
oscribe-cli = "oscribe.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/oscribe"]
